requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
distlib==0.3.4
et-xmlfile==1.1.0
fastapi==0.78.0
filelock==3.7.1
h11==0.13.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
openai==0.19.0
openpyxl==3.0.10
pandas==1.4.2
pandas-stubs==1.2.0.61
pip==21.2.2
platformdirs==2.5.2
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
requests==2.28.0
setuptools==61.2.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.14.1
wheel==0.37.1
wincertstore==0.2
PyJWT==2.6.0